add img compressor

+++ scaffold/src/main/java/com/vgaw/scaffold/ScaffoldSetting.java
@@ -1,5 +1,5 @@
 package com.vgaw.scaffold;
 
 public class ScaffoldSetting {
-    public static final int MAX_IMG_UPLOAD_SIZE = 6 * 1024 * 1024;
+    public static final long MAX_IMG_UPLOAD_SIZE = 6 * 1024 * 1024;
 }
